pyproject.toml
[project]
name = "rag-llamaindex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langfuse>=3.12.0",
    "langgraph>=1.0.7",
    "langgraph-prebuilt>=1.0.7",
    "llama-index>=0.14.13",
    "llama-index-llms-openai>=0.6.13",
    "llama-index-readers-file>=0.5.6",
    "llama-index-vector-stores-weaviate>=1.4.1",
    "pydantic>=2.0.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "weaviate-client>=4.19.2",
]
